Passing some ldr test, tests related with catching memory access out of bound hasn't been solved. In the process of fixing condition code failuer right now

+++ src/emulate/emulate_main.h
@@ -56,7 +56,7 @@ enum OpCode{
 };
 
 void emulate(struct EmulatorState *state,
-             uint32_t instructions[],
+             uint8_t instructions[],
              unsigned int instructions_l);
 
 #endif //EMULATE_EMULATE_MAIN_H
